Split constants and exceptions into separate modules
