+++ ariadne/__init__.py
@@ -1,16 +1,9 @@
 from .build_schema import build_schema_from_type_definitions
 from .executable_schema import make_executable_schema
-from .playground import PLAYGROUND_HTML
 from .resolvers import add_resolve_functions_to_schema, default_resolver, resolve_to
-from .wsgi_middleware import (
-    GraphQLMiddleware,
-    HttpError,
-    HttpBadRequestError,
-    HttpMethodNotAllowedError,
-)
+from .wsgi_middleware import GraphQLMiddleware
 
 __all__ = [
-    "PLAYGROUND_HTML",
     "GraphQLMiddleware",
     "add_resolve_functions_to_schema",
     "build_schema_from_type_definitions",
