+++ ariadne/constants.py
@@ -1,3 +1,13 @@
+DATA_TYPE_JSON = "application/json"
+
+CONTENT_TYPE_JSON = "application/json; charset=UTF-8"
+CONTENT_TYPE_TEXT_HTML = "text/html; charset=UTF-8"
+CONTENT_TYPE_TEXT_PLAIN = "text/plain; charset=UTF-8"
+
+HTTP_STATUS_200_OK = "200 OK"
+HTTP_STATUS_400_BAD_REQUEST = "400 Bad Request"
+HTTP_STATUS_405_METHOD_NOT_ALLOWED = "405 Method Not Allowed"
+
 PLAYGROUND_HTML = """
 <!DOCTYPE html>
 <html>
